Refactor reporting and billing services to use ReportDto. (#335)
* Refactor reporting and billing services to use ReportDto.

Updated the reporting structure by introducing a generic `ReportDto` to encapsulate data and size, replacing specific extensions in report DTOs. Refactored `generateBilling` to `generateBillingReport` for improved sorting, pagination, and clarity in the billing service. Added utility methods in `ReactAdminParams` for enhanced functionality.

* Refactor variable naming in getComparator method.

Renamed variables for clarity and consistency, changing "billing1" and "billing2" to "object1" and "object2". This improves readability and better reflects the generic nature of the method.

* "Remove @NotNull annotation from getComparator method"

---------

Co-authored-by: eazaretzky <esteban@entropyteam.com>